requirements.txt
# Core ML/AI dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Image processing
Pillow>=9.0.0

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Additional utilities
numpy>=1.21.0
